Carousol basics setup, re routed pages(input and output pages), Container basics setup
