+++ treehacks2024/src/App.tsx
@@ -1,45 +1,18 @@
 import { Button } from "@/components/ui/button";
-import { SignInButton, UserButton } from "@clerk/clerk-react";
-import {
-  Authenticated,
-  Unauthenticated,
-  useConvexAuth,
-  useMutation,
-  useQuery,
-} from "convex/react";
-import { UploadButton, UploadFileResponse } from "@xixixao/uploadstuff/react";
-import "@xixixao/uploadstuff/react/styles.css";
-import { api } from "../convex/_generated/api";
+import { SignInButton } from "@clerk/clerk-react";
+import { Authenticated, Unauthenticated, useConvexAuth } from "convex/react";
 import ShuffleHero from "./Shufflehero.jsx";
-import Container from "./main/Containter.tsx";
-import Navbar from "./main/Navbar.tsx";
-
-
+import Navbar from "./main/components/Navbar.tsx";
+import InputPage from "./main/pages/InputPage.tsx";
 
 export default function App() {
   console.log(useConvexAuth());
 
-  const generateUploadUrl = useMutation(api.files.generateUploadUrl);
-  const saveStorageId = useMutation(api.files.saveStorageId);
-  const saveAfterUpload = async (uploaded: UploadFileResponse[]) => {
-    await saveStorageId({
-      uploaded: { storageId: (uploaded[0].response as any).storageId },
-    });
-  };
-
   return (
     <main>
+      <Navbar />
       <Authenticated>
-        <SignedIn />
-        <UploadButton
-          uploadUrl={generateUploadUrl}
-          fileTypes={[".pdf", "image/*"]}
-          onUploadComplete={saveAfterUpload}
-          onUploadError={(error: unknown) => {
-            // Do something with the error.
-            alert(`ERROR! ${error}`);
-          }}
-        />
+        <InputPage />
       </Authenticated>
       <Unauthenticated>
         <ShuffleHero />
@@ -50,48 +23,5 @@ export default function App() {
         </div>
       </Unauthenticated>
     </main>
-  );
-}
-
-function SignedIn() {
-  return (
-    <>
-      <Navbar />
-      <Container></Container>
-      <p className="flex gap-4 items-center">
-        This is you:
-        <UserButton afterSignOutUrl="#" />
-      </p>
-
-      {/* <p>
-        Click the button below and open this page in another window - this data
-        is persisted in the Convex cloud database!
-      </p>
-
-      <p>
-        Edit{" "}
-        <code className="relative rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold">
-          convex/myFunctions.ts
-        </code>{" "}
-        to change your backend
-      </p>
-      <p>
-        Edit{" "}
-        <code className="relative rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold">
-          src/App.tsx
-        </code>{" "}
-        to change your frontend
-      </p>
-      <p>
-        Check out{" "}
-        <a
-          className="font-medium text-primary underline underline-offset-4"
-          target="_blank"
-          href="https://docs.convex.dev/home"
-        >
-          Convex docs
-        </a>
-      </p> */}
-    </>
   );
 }
