+++ treehacks2024/src/main/pages/OutputPage.tsx
@@ -1,0 +1,1 @@
+// shows up once submitted photo and response is generated.
